View 오류 수정
